feat(pmc): add SDMA usage metric and minor collector improvements
Add SDMA (System DMA) usage as a new GPU metric for PMC collection,
enabling monitoring of DMA engine utilization.

Key changes:
- Add sdma_usage metric (ID 14) with accessor, setter, and mask
- Add SDMA to user metric aliases and Perfetto track info
- Update metric validation pattern to accept "sdma_usage"
- Add pytest marker for SDMA tests

Additional fixes:
- Change get_use_rocpd/get_caching_perfetto to return bool by value
- Fix store_sample parameter order (enabled before supported)
- Use core/amd_smi.hpp wrapper instead of direct amdsmi.h include
- Change sample.device_id from size_t to uint32_t
- Add debug logging for perfetto processor creation

+++ projects/rocprofiler-systems/source/lib/rocprof-sys/library/pmc/collectors/gpu/sample.hpp
@@ -38,7 +38,7 @@ struct sample : trace_cache::cacheable_t
     {}
 
     enabled_metrics enabled_metric;
-    size_t          device_id;
+    uint32_t        device_id;
     uint64_t        timestamp;
     metrics         metric_values;
 };
